+++ react/src/js/components/Consonant/Cards/__tests__/HalfHeight.spec.js
@@ -118,7 +118,7 @@ describe(`Consonant/Card/${cardStyle}`, () => {
         const cardLogo = screen.queryByTestId('consonant-Card-logo');
         expect(cardLogo).toBeNull();
     });
-    
+
     test('should not render a card text', () => {
         renderCard({
             cardStyle,