+++ react/src/js/components/Consonant/Cards/__tests__/TextCard.spec.js
@@ -73,7 +73,7 @@ describe(`Consonant/Card/${cardStyle}`, () => {
         const labelElement = screen.queryByTestId('consonant-Card-text');
         expect(labelElement).not.toBeNull();
     });
-    
+
     test('should be able to render a card footer', () => {
         renderCard({
             cardStyle,